Refactor : 여정 조회 api 함수 분리
